Remove single-obsdataset condition in colocate_gridded_ungridded (to allow colocation with UngriddedData objects containing multiple datasets); move beta method correc_model_stp_coldata from colcoation.py into new module not_finalised.py
